// frontend/src/types.ts
export interface IrrigationRequest {
  state: string;
  district: string;
  crop: string;
  area_ha?: number;
  start_date?: string | null;
  pump_hp?: number;
}
// ...
export interface UserFarmProfile {
  // Place & Area are compulsory
  state: string;
  district: string;
  area_ha: number;
  // All other fields are optional
  farmer_name?: string;
  user_name?: string;
  mobile?: string;
  user_phone?: string;
  village?: string;
  crop?: string;
  soil_type?: string;
  irrigation_source?: string;
  sowing_date?: string;
  pump_hp?: number;
  user_email?: string;
}